# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Clarity B2B API",
    "version": "1.0.0",
    "description": "API documentation for the Clarity Supply Chain Finance backend. This covers the Invoice Pipeline and Supplier Health Analytics modules."
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/confirmations/pending": {
      "get": {
        "summary": "Get pending confirmations",
        "description": "Get a list of invoices that require buyer confirmation.",
        "parameters": [
          {
            "name": "buyer",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the buyer"
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON array of pending invoices"
          }
        }
      }
    },
    "/api/confirmations/history": {
      "get": {
        "summary": "Get confirmation history",
        "description": "Get a list of invoices that have already been confirmed or disputed by the buyer.",
        "parameters": [
          {
            "name": "buyer",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the buyer"
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON array of processed invoices"
          }
        }
      }
    },
    "/api/confirmations/{id}/confirm": {
      "post": {
        "summary": "Confirm an invoice",
        "description": "Confirm an invoice as a buyer, moving it to 'Buyer Confirmed' status.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "buyer_name": {
                    "type": "string"
                  },
                  "acknowledgment_text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice confirmed successfully"
          },
          "404": {
            "description": "Invoice not found or unauthorized"
          }
        }
      }
    },
...
  }
}